avoid redundant propagation for tuples_in/2
Currently limited to tuples_in/2. Enabling it for other constraints
resulted in slowdown of several benchmarks (711.pl, nut.pl) while
improving others (golomb.pl, mhex.pl). As soon as the other
constraints are formulated in such a way that it does not slow them
down, this should be enabled for all of them.