ADDED: Syntax closer to ISO with set_prolog_flag(iso,true).
See Prolog flag iso for a detailed description.

Examples for new errors: f(,,a). [|]. f(a:-b). X == -, t.

The change maintains the following property: If a text parses without
an error with current_prolog_flag(iso,true), it will parse to the very
same term with current_prolog_flag(iso,false).

There is much more to this, e.g. in 6.3.4.3 Operators it is said:
"There shall not be an infix and a postfix operator with the same
name".  On the other hand, the NOTE in 5.5.1 Syntax extensions
mentions this "as an implementation specific feature ...".  It seems
almost all implementations do support this. TBD