FIXED: variant/2 take 42 ...  Kuniaki Mukai, Ulrich Neumerkel.
After some Ulrich found some clever counter examples for take 41, I had
some more discussion with Kuniaki.  I've now implemented a rather naive
term-walk that unfortunately does not optimize for last arguments.  The
implementation avoids the C-stack, so this shouldn't be too bad.  It no
longer exploits sharing, but it appears there is not much to exploit if
it comes to sharing.

There are options to reduce the stack-usage (which is now 5 words per
node, where the node-count depends on the recursion depth of the terms.
This can quite easily be reduced to 3 words by processing the arguments
last-to-first and storing the saved mark-bits with the node-pointers.

Avoiding dependency on the depth for the last argument might be feasible,
either using Kuniaki's approach (unsure) or using the old term-linking
approach and revert to the new appraoch only if we detect that the
term-link appraoch may yield a false positive.