FIXED: Various issues around the function random/1 if implemented by GMP
GMP-based random/1 uses a state per thread.  On new threads, this was
initialised from /dev/urandom or /dev/random or time().  The initialization
based on the devices could not use all bits if it was broken into multiple
read calls (possible, especially with /dev/urandom if entropy is
insufficient).

The fallback initialised on the wall-time in seconds using time(). That
works out really poor if actions that need a random number start a new
thread.

The fall-back was always used on Windows.  The current implementation is
modified by Keri Harris to use CryptGenRandom(), which is supposed to be
similar to Linux /dev/(u)random.

The fall-back now uses WallTime() from os/pl-os.c, which returns the wall
time at the highest resolution supported by the OS.