ADDED: prolog_trace_interception/4 action 'skip'.
This fixes skip over a cut in the GUI tracer. This patch also updates
the documentation of the action return-argument of this predicate.
Finally, this patch fixes the use of a long int in PL_get_frame(), which
can crash the debugger on Win64 systems when using a really big (>16Gb
local stack).