FIXED: No longer treat Windows SOCKET handles as file numbers.
Using these handles with non-winsock APIs crashes the application and
the types do not match. There is a new function Swinsock() that
retrieves an object of type SOCKET (unsigned int, but this assumption
is no longer used).  This is only implemented in the Windows version.

This patch fixes crashes to do file-operations operating on SOCKET
type objects.  It also simplifies and improves the implementation
of wait_for_input/3, notably with many open file descriptors.