ADDED: Trying to get fork() work from a threaded application. Getting the whole thing working is probably hard, but it might be possible to guarantee a working fork from a threaded SWI-Prolog under some conditions.
The idea of this patch is to use pthread_atfork() to re-initialise
the thread administration to reflect the new situation.  E.g., the
only thread is thread 1 (but the thread-id of the forked thread
may be different) and all other threads have gone.

Currently not enabled by default.  Add -DO_ATFORK to enable this.