PORT: 0004-swipl-ld-prog-out.patch.  Kery Haris.
When configure scripts test if swipl-ld-build.exe work, they compile
conftest.c and test whether the following files are executable: {a.out,
conftest.exe, conftest, a.exe, a_out.exe, b.out, conftest.c}.
swipl-ld-build.exe defaults to plout.exe. If we are on Windows and
HOST_TOOLCHAIN_MSC is not set then we should default to a.out.