PORT: More absolute path issues
Unix absolute paths cannot be handed to the MinGW version because the
same path has a different meaning there. There are two ways out: (1)
avoid absolute paths or (2) use winepath -w to convert.

We opt for (1), but this requires changing directory before running the
commands in some cases.  Therefore, we adapted swipl.sh to accept a -C
option like make.