FIXED: Race condition updating hash-indices.  Matt Lilley.
hashDefinition() should do the detection of an old table while
holding the lock.  The computation is still outside.  This allows
for static predicates (who share locks) to be indexed in parallel.
In the unhappy event that two threads index at the same time, one
of the indexes will be discarded through the normal scenario of
GC-ing indexes.