ENHANCED: Speedup lookup from file to module.
The previous implementation enumerated all modules. The new
implementation maintains a list of modules associated to a file to
answer this query without search.

The interface of '$current_module'/2 for mode (-,+) has become a bit
awkward. See docs. This is hard to avoid because we cannot keep a lock
on the linked list, which makes backtracking in C rather complicated.