FIX: thread_get_message/3 relative time timeout/1 option now works correctly, ADDED: thread_get_message/3 now accepts absolute time deadline/1 option.
The correct implementation of the timeout option of thread_get_message/3
was much easier to get by working with an absolute time deadline
throughout. This absolute time deadline is exposed at the interface
using the deadline/1 option, making strongly-timed (where the times of
occurence of computations are part of the semantics of the program and
are as important as other facets such as values and control flow to the
correctness of the implementation) code easier to write at the Prolog level.