FIXED: Handle goal_expansion/2 for local meta-predicates
Goal expansion was not performed on <goal>  in the code below. Note that
the current behaviour  is  still   somewhat  troublesome  because  later
(auto-)importing relies on whether or not   a  predicate is defined. The
new  code  considers  a  goal  candidate  for   expansion  if  it  is  a
meta-predicate, whether defined or not. Maybe we should also immediately
import?

:- meta_predicate local(0).

test :- local(<goal>).

local(Goal) :- Goal.