ENHANCED: Reimplemented library(prolog_metainference) using attributed variables for annotating.  This allows for richer annotations and avoids search in lists of variables.
Currently, alias analysis is left out. This used to propose `0' for
arguments that were aliased to meta-arguments, but these should be
annotated with `:' rather than `0'. It is not really clear how to
proceed. We can use simple alias analysis like =/2 that will give
precise and meaningful analysis. We can also look for some patterns like
A =.. List, <something with list>, B =.. NewList, call(B), but there are
many variations within <something with list>.  Basically, all we want to
know is how the length of NewList compares to List.  Similar stories with
functor/3.