CLEANUP: Stop analysing internals of libraries for missing meta-predicate declarations.
Here, we have two options. Either fix the libraries or stop the
warnings. Fixing the libraries looks attractive, but the performance of
include/3 drops with almost 10% if we declare the internal helper as a
meta-predicate. Unfortunately, this is rather difficult to fix in the
SWI-Prolog design for meta-predicates, where the qualification happens
in the predicate supervisor instead of in the caller.