CLEANUP: Various issues in the Makefile, introducing .PHONY targets.
Done searching why the Windows build often ends with this:

==
rm -f swipl.prc
./swipl.sh -O -o swipl.prc -b ../boot/init.pl
./swipl.sh -x swipl.prc \
			-f none -F none \
			-g make_library_index\(\'../library\'\) \
			-t halt
[PROLOG SYSTEM ERROR:  Thread 1
	trapUndefined(): undefined: system:$find_library/5

PROLOG STACK:
==

But, although this only seems to happen with parallel builds, it
doesn't seem to be a concurrency issue.  The first call does create
swipl.prc, but it prints no messages and the state only contains
init.pl, not the other boot files.