ADDED: packs: allow info-term branch(Branch).
Prolog Packs currently assume that the Git master branch contains the
pack. This need not be the case however, since the master branch could
be used for continuous development.

This patch adds support for setting `branch/1` in =pack.pl= to e.g.
branch(release) or branch_name(prolog). Prolog Pack makes
sure to checkout to this branch after cloning the repositoty.