ENHANCED: CLP(FD): First preparations to support monotonic constraints.                    Ulrich Neumerkel (i3a#313). Support X? notation which                    constrains X to integers in CLP(FD) expressions.
A current declarative shortcoming of CLP(FD) is that it commits to
interpreting variables that occur in a query as finite domain
variables (= integers) instead of logical variables. For example:

   ?- X #= 3.
   %@ X = 3. (no further solutions)

But X might as well stand for a different expression, for example:

   ?- X = (Y + Z), X #= 3.
   %@ X = Y+Z,
   %@ Y+Z#=3.

Note that exchanging the goals yields a different result:

   ?- X #= 3, X = (Y + Z).
   %@ false.

As another example, we have:

   ?- #\ 2.
   %@ false.

because the only solution of

   ?- #\ X.

is currently:

   %@ X = 0.

whereas declaratively, X = (5 #= 3) is for example also a correct
answer, and indeed we have:

   ?- X = (5 #= 3), #\ X.
   %@ X = (5#=3).

To fix this, the (?)/1 syntax is now made available to explicitly
denote finite domain variables in CLP(FD) expressions. When you add:

   :- op(5, yf, ?).

to your programs, you can now write:

   ?- X? #= 3.
   %@ X = 3.

The above example then yields:

   ?- X = (Y + Z), X? #= 3.
   %@ ERROR: Type error: `integer' expected, found `_G759+_G760'

because "X?" constrains X to integers. This renders your CLP(FD)
programs monotonic: adding additional goals cannot yield additional
answers.

As another example, consider again:

   ?- #\ X? .
   %@ X = 0.

and contrast it with:

   ?- X = (5 #= 3), #\ X? .
   %@ ERROR: Type error: `integer' expected, found `5#=3'

The convenient and declaratively problematic syntax

   ?- X #= 3.
   %@ X = 3.

is currently still available. To enforce a more declarative coding
style, such a query should throw an instantiation error because too
little information is known about X to answer correctly. Add (?)/1 to
constrain it to integers, and exclude other CLP(FD) expressions.