Modernize README.cygwin
Hi,

the below patch revamps the README.cygwin file slightly to keep up with
reality.

Note:  It already contains a few filename and path changes which are not
yet in the repo.  I'm working on the required patches and will be able
to send them today or tomorrow, I hope.

Jan, in terms of the Drawback section, I retained it, just rephrased
a bit to be more neutral.

Thanks,
Corinna