Simplify. Please see below for more information.
First of all, the else-branch can never be encountered in the current
version. This is because we are only considering the *root* branching
variable after unifications, and all other variables have greater
index. Therefore, the OVI < VI branch never applies.

However, suppose we generalize Vars in satisfiable_bdd/2 to all
branching variables that actually occur in the BDD. (As an aside, this
requires a modification in nodes_variables_//1 to not accidentally
call put_visited/1 on instantiated variables, leading to an
uninstantiation error. Suppose in the following that we change
nodes_variables_//1 to simply ignore instantiated variables.)

Then, it *is* possible that the else-branch is encountered. Example:

   ?- sat(A+ (B+ (C+D))), [A,C]=[0,0].

*However*, it does not make a difference, because when encountering
instantiated variables, then var_index/2 in single_truth_value/4
fails, and thus does not yield a false positive even in cases where
index_skipped/2 would (as in: without the else-branch) falsely
indicate that the index was not skipped.

Such considerations are especially (or even: only) necessary for the
case of simultaneous unifications.

I have found the following program useful to test simultaneous
unifications, and to compare them with a sequence of single ones:

    :- use_module(library(clpb)).

    exp(T, Vs0, Vs) --> { select(T, Vs0, Vs) }.
    exp(~T, Vs0, Vs) --> { select(T, Vs0, Vs) }.
    exp(A+B, Vs0, Vs) --> [_], exp(A, Vs0, Vs1), exp(B, Vs1, Vs).
    exp(A*B, Vs0, Vs) --> [_], exp(A, Vs0, Vs1), exp(B, Vs1, Vs).

    exp(T, Vs) :- phrase(exp(T,Vs,_), Vs, _).

    list_simul_assignments(Ls0, Assignments) :-
            permutation(Ls0, Ls),
            phrase(lists(Parts), Ls),
            maplist(simul_assignment, Parts, Assignments).

    simul_assignment(Ls, Ls = Bs) :-
            same_length(Ls, Bs),
            maplist(boolean, Bs).

    lists([]) --> [].
    lists([L|Ls]) --> olist(L), lists(Ls).

    olist([L|Ls]) --> [L], list(Ls).

    list([]) --> [].
    list([L|Ls]) --> [L], list(Ls).

    boolean(0).
    boolean(1).

    run :- length(_, N), portray_clause(N), run(N).

    run(N) :-
            length(Vs, N),
            exp(T, Vs),
            term_variables(T, Vs),
            list_simul_assignments(Vs, Es),
            Goal = (sat(T), maplist(call, Es)),
            \+ \+ catch(Goal, E, writeln(E-Goal)),
            phrase(separate_assignments(Es), Es1),
            Goal1 = (sat(T),maplist(call, Es1)),
            findall(Vs, (Goal,labeling(Vs)), Sols),
            findall(Vs, (Goal1,labeling(Vs)), Sols1),
            (   Sols == Sols1 -> true
            ;   throw(Goal-Goal1)
            ),
            false.

    separate_assignments([]) --> [].
    separate_assignments([As=Bs|Rest]) -->
            separate_(As, Bs),
            separate_assignments(Rest).

    separate_([], []) --> [].
    separate_([X|Xs], [Y|Ys]) --> [X=Y], separate_(Xs, Ys).

Example:

   ?- run.
   %@ 0.
   %@ 1.
   %@ 2.
   %@ 3.
   %@ etc.

For example, if we change the else-branch under consideration to raise
an exception, say 'branch', in order to see whether it is executed,
and apply the indicated modifications to satisfiable_bdd/2 and
nodes_variables_//1, then we get:

    %?- run.
    %@ 0.
    %@ 1.
    %@ 2.
    %@ 3.
    %@ 4.
    %@ branch- (sat(_G94+ (_G97+ (_G100+_G103))),maplist(call,[[_G94,_G100]=[0,0],[_G97]=[0],[_G103]=[0]]))
    %@ ERROR: Unhandled exception: branch

In addition, here is one more example that shows why the current
generalised reasoning is so desirable. Previously, we had for example:

   %?- sat(X+Y+Z), sat(Y=:=0).
   %@ sat(X=:=X),
   %@ node(9)- (X->node(7);node(8)),
   %@ node(8)- (Y->false;node(3)),
   %@ node(3)- (Z->true;false),
   %@ node(7)- (Y->false;true),
   %@ sat(Y=:=Y),
   %@ sat(Z=:=Z).

This means that even quite literal "unifications" were not propagated,
simply because the variable was not at the root.

In contrast, we get with more recent versions:

   %?- sat(X+Y+Z), sat(Y=:=0).
   %@ Y = 0,
   %@ sat(X=:=X),
   %@ node(10)- (X->true;node(3)),
   %@ node(3)- (Z->true;false),
   %@ sat(Z=:=Z).

This makes =:= and = quite uniform when only a single variable is
involved, and this is what one would generally expect.