mention the need for projecting residual constraints to query variables
To see how useful this would be, consider the following example from
Ulf Nilsson's excellent lecture slides, implementing a XOR circuit:

   :- use_module(library(clpb)).

   n_mos(S,G,D) :- sat(S*G =:= D*G).

   p_mos(S,G,D) :- sat(S* ~G =:= D* ~G).

   xor(X, Y, Z) :-
           p_mos(Tmp, 1, X),
           n_mos(0, Tmp, X),
           p_mos(Z, X, Y),
           n_mos(Z, Tmp, Y),
           n_mos(Z, Y, Tmp),
           p_mos(Z, Y, X).

Now, we can use CLP(B) to see that this really implements XOR:

   ?- xor(X, Y, Z), taut(Z =:= X#Y, T).
   %@ T = 1,
   %@ etc.

However, the residual goals contain a variable that does not occur in
the query:

   %@ sat(X=:=X),
   %@ sat(1#X#X*Y*Z#X*Y*Z*_G405#X*Y*_G405#X*Z*_G405#Y#Y*Z#Y*Z*_G405#Z),
   %@ sat(_G405=:=_G405),
   %@ sat(Z=:=Z),
   %@ sat(Y=:=Y).

We can manually quantify this variable existentially as follows:

   ?- sat(_G405^(1#X#X*Y*Z#X*Y*Z*_G405#X*Y*_G405#X*Z*_G405#Y#Y*Z#Y*Z*_G405#Z)).

and obtain:

   %@ sat(_G405=:=_G405),
   %@ sat(1#X#Y#Z),
   %@ sat(X=:=X),
   %@ sat(Y=:=Y),
   %@ sat(Z=:=Z).

This shows that the circuit indeed implements Z =:= X#Y.

This existential quantification of variables that do not occur in the
query should be done implicitly by the Prolog engine in the future. In
the case of CLP(B), this is very easy to add, because existential
quantification of CLP(B) variables is already implemented and only
needs to be used when projecting attributes to residual goals.

Another example:

   ?- xor(X, Y, 0).
   %@ sat(X=:=X),
   %@ sat(1#X#X*Y*_G312#Y),
   %@ sat(_G312=:=_G312),
   %@ sat(Y=:=Y).

Again, we currently have to add the existential quantification manually:

   ?- sat(_G312^(1#X#X*Y*_G312#Y)).

and obtain:

   %@ sat(_G312=:=_G312),
   %@ sat(1#X#Y),
   %@ sat(X=:=X),
   %@ sat(Y=:=Y).

This shows that when Z is 0 in the circuit, X is Y.