ENHANCED: CLP(B): Faster domain and aliasing consistency.
We now determine all variables that are not skipped in any branch of
the BDD in essentially a single pass.

This yields significant performance improvements:

   %?- comparison.
   %@ euler172(6)         1.689     2.149
   %@ cycle100i           1.022     1.000
   %@ queens6             1.129     1.457
   %@ cycle100k           1.040     1.000
   %@ dominos7            1.062     1.000
   %@ queens7             1.205     1.754
   %@ euler172(8)         2.084     2.474
   %@ dominos8            1.051     1.000
   %@ euler172(10)        1.770     2.721
   %@ false.

This is the comparison with the latest released version that didn't
have aliasing consistency at all:

   %?- comparison.
   %@ euler172(6)         0.779     0.846
   %@ cycle100i           1.120     1.000
   %@ queens6             0.343     0.553
   %@ cycle100k           1.259     1.000
   %@ dominos7            1.182     1.000
   %@ queens7             0.265     0.572
   %@ euler172(8)         0.865     0.841
   %@ dominos8            1.340     1.000
   %@ euler172(10)        0.698     0.855
   %@ false.

So, the recent changes (most significantly: said aliasing consistency)
have a big performance impact in the N-queens examples, and are
otherwise quite acceptable, when taking into account the very nice
resulting algebraic properties. See for yourself, for example:

   ?- sat(A*B + ~A * ~B).
   %@ A = B,
   %@ sat(B=:=B).