FIXED: CLP(FD): Apply goal expansion only if the CLP(FD) predicates are visible in the load context. Jan Wielemaker.
For example, suppose we have a.pl:

   :- module(a, []).
   :- use_module(library(clpfd)).

and b.pl:

   :- module(b, [test/0]).

   test :- in(_, _).

   in(x, y).

Previously, test/0 would read:

   test :- clpfd:clpfd_in(X, Y).

This was obviously not intended, and is now corrected.

The new implementation (many thanks Jan!) even deals with:

   :- use_module(library(clpfd)).

   in(x, y).

   ...,
   in(X, Y).

i.e., if in/2 is defined before using. It will not deal with the case
where in/2 is defined *after* being called, because the current
infrastructure needs some extensions before this can be implemented.