ENHANCED: CLP(B): Make residual goals work for recorded constraints.
Not doing this leads to the following problem in SWISH, which uses the
database to record results:

    ?- sat(card([1], [A+B])), recorda(s, [A+B]).
    %@ sat(A=\=A*B#B).

    ?- recorded(s, T).
    %@ false.

This is because when the constraints are fetched and the constraints
are projected to residual goals, CLP(B) would again use IDs 0,1,...
for variables (because the counter is reset), but these IDs are
already used in the recorded attributes. Thus, existing IDs are
inadvertently reused for new variables.

Ideally, SWISH should simply obtain the residual goals with
copy_term/3, and then work on the residual consraints alone.

With this change though, it also works now:

    ?- sat(card([1], [A+B])), recorda(s, [A+B]).
    %@ sat(A=\=A*B#B).

    ?- recorded(s, T).
    %@ T = [_G192+_G193],
    %@ sat(_G192=\=_G192*_G193#_G193).