DOC: library(assoc): Prepare for using PlDoc directly in the manual.
This change aligns the PlDoc comments with the existing reference
documentation in man/assoc.doc. It incorporates everything that was
missing, fixes a few mistakes, and names the arguments consistently:
Arg0 --> Arg is now consistently used whereas previously both ArgOut
and NewArg were used to denote output arguments.

This sets the stage for using the PlDoc reference documentation
directly in the manual, allowing us to get rid of the .doc version.