ENHANCED: (#=)/2 now automatically uses powm/3 in many relevant cases.
It is the job of the Prolog environment, not of application programmers,
to use fast low-level predicates in cases of practical relevance. Ideally,
programmers only need to use (#=)/2 to express equality over integers,
without having to care about ordering, instantiation, special cases etc.

For example, consider commit b380cad24cbe5102a549bfb8f7406e34a2b99e1a
in the SSL package, where Pow #= Signature^Exponent mod Modulus is used
to verify a certificate's signature with the CLP(FD) constraint (#=)/2.

If the arguments are sufficiently instantiated, this constraint is now
*automatically* compiled to Pow is powm(Signature,Exponent,Modulus),
making the query run ca. 30 000 times faster. This also works if we write
the constraint as Signature^Exponent mod Modulus #= Pow, which may be
more natural to Prolog programmers with a strong mathematical background.

Integer constraints of the form A^B mod M also occur in DH key exchange,
which is used to ensure forward secrecy in (HTTPS) web applications.