Add docs, remove rb_lookupall/3 and fix rb_in/3
These changes are aimed at clarifying the status of variables in keys
and duplicate keys. After discussions on the mailing list, we
established that is is ok for keys to contain variables, in which case,
key matching is based on (==)/2 equivalence of terms (and not on
unifiability, subsumption or structural equality).

It also seems to be the case that a tree with more than one entry for a
given key is an invalid tree (and is rejected by is_rbtree/1).

Taking these together, we can conclude that the old rb_lookupall/3
was redundant, since there should only ever be one value for a given
key whether or not that key contains variables. Instead, rb_lookup/3
should be used. I've taken the possibly drastic step of removing
rb_lookupall/3 completely, but another option would be to leave it in
but deprecated and with dire warnings.

Also, the old rb_in/3 was diverting to lookup/3 for nonvariable keys.
This was giving incorrect behaviour for partially instantiated keys,
since the tree may contain several keys that unify with the given term.
My initial fix was to replace the nonvar(Key) with a \+ground(Test).
However, further consideration has lead me to remove that clause of
rb_in/3 entirely. The reasoning behind this is:
   (a) if you want fast lookup, use rb_lookup/3;
   (b) making rb_in/3 check the key argument with ground/1 forces it to
       examine the whole term and removing that clause makes rb_in/3
       cleaner;
   (c) the rest of the rbtrees interface is relies on (==)/2 equality
       between keys, where as the nondeterministic mode of rb_in/3
       relies on unification of keys. This is fine, but mixing both
       meanings in one predicate depending on the instantiation state
       of the key muddies the semantics, relies on extra-logical
       which lead (as we have seen) to a bug, and generally feels not
       quite right.
In further support of this change, this is exactly what
assoc:gen_assoc/3 does, leaving get_assoc/3 the job of doing O(log N)
lookups. Also, assoc:ord_list_to_assoc also prevents the creation of
duplicate keys.