FIXED: CLP(FD): Overeager goal expansion. Julien Cumin.
This mistake was introduced in 038d86b08bfecc0715b4548a6ec3f2d91c20b658.

A new binding *must not* be dragged outside of disjunctions, since the
code may look for example like this:

        i(X) :-
            (   X #= 3
            ;   X #= 4
            ).

In fact, this would previously not even compile, and instead raise:

    put_attr/3: Uninstantiated argument expected, found 3 (1-st argument)

already during goal expansion.

This commit fixes this issue, and still rewrites CLP(FD) expressions
as far as possible already at compilation time.

For example:

    n(X) :- X #= 1+3.

This is now compiled to (note that 1+3 is evaluated to 4):

    ?- listing(n/1).
    %@ n(A) :-
    %@      (   integer(A)
    %@      ->  A=:=4
    %@      ;   var(A)
    %@      ->  A=4
    %@      ;   B=4,
    %@          clpfd:clpfd_equal(A, B)
    %@      ).

Ideally, it should be compiled to:

    n(4).