Handling of new answers in mode directed tabling
Hi Jan,
I am proposing this change because when an answer is updated by mode directed tabling, it must be added to the worklist to be considered for further processing. This solves the problem I was writing about. Unfortunately it adds redundant answers, so I guess one should check that the updated answer is different from the one already in the trie but I don't know how to do it.

Fabrizio