DOC: Improve description of the new mitigate_spectre flag.
In particular, note that the ability to run subsets of Prolog code,
let alone arbitrary Prolog code, is not a necessary condition for
Spectre attacks. It suffices that sensitive data is speculatively
accessed, and that this data can be leaked. Also trusted code may be
vulnerable to this if it contains branches that are dependent on
untrusted data.

Decreasing the resolution of timers does _not_ reliably prevent
Spectre attacks. Full mitigation may require compiler support to
disable speculative access to sensitive data.

See https://github.com/SWI-Prolog/roadmap/issues/54 .